+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.16.0]
+
 ### Added
 
 - `variable_files` plan keyword for reading variables from JSON or YAML file.
@@ -51,7 +53,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for executing multiple plans at once.
 - Support for running plans in parallel.
 
-[unreleased]: https://github.com/kangasta/yaml_requests/compare/v0.15.0...HEAD
+[unreleased]: https://github.com/kangasta/yaml_requests/compare/v0.16.0...HEAD
+[0.16.0]: https://github.com/kangasta/yaml_requests/compare/v0.15.0...v0.16.0
 [0.15.0]: https://github.com/kangasta/yaml_requests/compare/v0.14.0...v0.15.0
 [0.14.0]: https://github.com/kangasta/yaml_requests/compare/v0.13.0...v0.14.0
 [0.13.0]: https://github.com/kangasta/yaml_requests/compare/v0.12.0...v0.13.0